Four things that were quietly not working, and the awards that never happened
ACADEMY INTAKE CRASHED THE SEASON ROLLOVER. It allocated a prospect id with an unbounded
`SELECT MAX(id)+1 FROM players` read through Convert.ToInt32. Measured on the real career
that query returns 50,000,004,410, so it threw OverflowException — inside rollover, after
relegation, ageing, loan returns and the CPU market had already been written. The career was
left half-rolled and the academy empty forever. Allocation is now bounded to the academy's own
band [30m, 40m), read as a long, and skips an intake rather than minting ids that collide with
the imported, curated or generated namespaces.

UNDO PAID FOR THE WEEK TWICE. UndoLastResult deleted cond_applied_{season}_{md} and
cond_cup_{...} along with the result. Those two rows are the only thing gating the weekly pass,
so undoing a mistyped score and re-entering it paid a full week of wages twice, took gate
receipts twice, handed the bank a second loan instalment, ran the training ground twice and
re-rolled every injury. Deleting the guard never "reset" the week — it doubled it. The guards
stay; the message now says plainly what an undo does not reverse, because the Office prints
that line verbatim.

THE OFFICE FROZE THE MOMENT YOU OPENED IT. Twelve gauges — board, morale, fans, balance,
expectation, ELO, tier, the chairman's line — were get-only properties read once in the
constructor. Record a result and the report above them updated while they went on showing the
club as it was when you arrived.

THE ASSISTANT'S TEAM SELECTION WAS A DECORATION. delegate_xi was the one delegation with no
body anywhere: the checkbox wrote a flag, the flag had a reader, and nothing acted on it. He
now picks, persists the side through squad_members.slot (so it reaches the game), and writes
you a letter naming who came in and who dropped out — only when the personnel actually change.
When you have picked by hand he stands aside and says so once, rather than silently ignoring
a ticked box.

AND, FOUND WHILE WRITING THE REGRESSION FOR IT: the end-of-season awards have never once been
recorded, in any career. honours.team_id declared REFERENCES teams(id), but an award row holds
a PLAYER id — and MasterDb opens with PRAGMA foreign_keys = ON. So the insert threw every
season, straight into the "the gala never blocks rollover" catch: no Player of the Season in
the Roll of Honour, and no awards letter at all, because PostInbox sits after the throw. The
column is polymorphic by design and now carries no key; existing databases are repaired by a
table rebuild in Migrate.

BOARD CONFIDENCE IS ONE NUMBER AGAIN. Session.Board is built once in the constructor, and
ApplyCareerAfterResult writes over board_conf, so a verdict from the season review, the
mid-season review or a refused budget ask moved a figure nothing read back — and the next
kick-off erased it. Swings now live in their own meta row, BoardConfidenceNow re-reads the
composite every time, and the six places that mattered ask for it: the Office gauge, the
post-match swing line, the Board screen, the trend chart — and the SACK, which had been
judging you on the results-only half.

Also: the archive's Roll of Honour reads its holder id as a long (same truncation, the read
side), the youth layer heals itself instead of trusting a completion flag, and the Academy
screen's youth tab gained portraits, a readiness read, a real empty state, a distinct failure
state and the missing U18 → U21 rung.

Proof: ActionsSmoke 44/44 against a copy of the career, with new regressions for the board
read-through, the undo guards, ownership of a player in your own U21s, and a player-sized
holder id surviving the archive round-trip. That last check is what turned the awards bug up.

+++ src/ML.App/SessionHistory.cs
@@ -118,10 +118,15 @@ public sealed partial class Session
     /// the team column, so a caller wiring a right-click menu has to know which it is before
     /// it offers to view a squad that does not exist.
     /// </summary>
-    public IReadOnlyList<(string Competition, string Holder, int HolderId, bool HolderIsPlayer)>
+    public IReadOnlyList<(string Competition, string Holder, long HolderId, bool HolderIsPlayer)>
         HonoursInWithHolders(int seasonId)
     {
-        var rows = new List<(string, string, int, bool)>();
+        // The holder id is a LONG. A 'pots' row stores a player id, and player ids in this world
+        // run past Int32 (the academy band alone reaches 50,000,005,004). GetInt32 on an int64
+        // column is an unchecked cast in Microsoft.Data.Sqlite, so it came back truncated and
+        // often negative: the archive showed Player of the Season with a blank name and its
+        // right-click menu refused to open. Club ids narrow at the call, where they're safe.
+        var rows = new List<(string, string, long, bool)>();
         using var cmd = Db.Connection.CreateCommand();
         cmd.CommandText = "SELECT competition, team_id FROM honours WHERE season_id=$s ORDER BY competition";
         cmd.Parameters.AddWithValue("$s", seasonId);
@@ -139,8 +144,8 @@ public sealed partial class Session
                 _ => CupName,
             };
             var isPlayer = raw == "pots";
-            var id = r.GetInt32(1);
-            rows.Add((comp, isPlayer ? PlayerNameOf(id) : TeamName(id), id, isPlayer));
+            var id = r.GetInt64(1);
+            rows.Add((comp, isPlayer ? PlayerNameOf(id) : TeamName((int)id), id, isPlayer));
         }
         return rows;
     }
